moved web api manager in a separate file
